Moving to Modal and fixing components
